Drop gen access & merge with engine
